Fix bug introduced by Versionista page view changing its structure
[Fixes Issue #8]
- The bug was introduced because Versionista changed
the structure of its 'page view' page. Specifically,
Versionista introduced an extra `<td>` element
in each row of the table of page versions. That
caused the prior xpath to fail to detect where
the comparison links were on the page. The revised
xpath looks for the first `<td>` that has at least
one `<a>` within it, and then looks for child `<a>`
links within that `<td>`, which solves the problem.